src/requirements.txt
rich
openai==0.28.0
google-api-python-client
youtube-transcript-api
google-auth-oauthlib
google-auth
selenium
pinecone-client
numpy
python-dotenv
google-cloud-firestore
aiohttp
pytest
firebase-admin
